Initial AI implementation
AI model created. No vision, but it looks like it will be needed after all. I am combining this as I will break the code by changing tensor dimensions.

Models has the car neural network. CarGameEnv is where we create the custom gym env and LearningSimulation is where the training is done.

Also modified main for the scoring system, then had to adjust UI because of it.